pyproject.toml
[tool.poetry]
name = "employeesmanagment"
version = "0.1.0"
description = ""
authors = ["AveGus(Eldar) <garipov.eldar200@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
django-filter = "^24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
